Learning about functions
Two functions where added
One test the length of a name and will change it depending on that length. black for less than 5 white for more than 10 and in between is a shade of gray

The other function takes two random numbers and adds them together. if the number is positive it will display green. If it is negative it will display red.